minor bug fixes

+++ src/opentyr.h
@@ -59,8 +59,8 @@ extern const char *opentyrian_str;
 
 char *strnztcpy( char *to, char *from, size_t count );
 
-int efread( void *buffer, size_t size, size_t num, FILE *stream );
-int efwrite( void *buffer, size_t size, size_t num, FILE *stream );
+size_t efread( void *buffer, size_t size, size_t num, FILE *stream );
+size_t efwrite( void *buffer, size_t size, size_t num, FILE *stream );
 
 void opentyrian_menu( void );
